Move Signal related code to Signal folder
Add project file references to all Signal headers. Update NAS2D.h to
include all Signal headers. Update unit test project paths.
