+++ NAS2D/Signal/Connection.h
