+++ NAS2D/Signal/Delegate.h
