+++ NAS2D/Signal/Signal.h
