+++ NAS2D/Signal/SignalSource.h
