+++ test/Signal/Delegate.test.cpp
@@ -1,4 +1,4 @@
-#include "NAS2D/Delegate.h"
+#include "NAS2D/Signal/Delegate.h"
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
