+++ test/Signal/Signal.test.cpp
@@ -1,4 +1,4 @@
-#include "NAS2D/Signal.h"
+#include "NAS2D/Signal/Signal.h"
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
